+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Devop
+
+- Contract creation [#912](https://github.com/EthVM/EthVM/pull/912)
+
 
 ### Release v1.0.3-hotfix.1